Factor out Tabs file
